Multi_Agent_Research_System/requirements.txt
langgraph
langchain
langchain-community
beautifulsoup4
faiss-cpu
numpy
tavily-python
python-dotenv
pydantic
langchain-google-genai 